+++ HISTORY.rst
@@ -67,11 +67,17 @@ Unreleased Changes (3.9)
     "333" leading to high export spikes in some pixels.
 
 ..
-Unreleased Changes
+..
+  Unreleased Changes
+  ------------------
+
+3.8.8 (2020-09-04)
 ------------------
 * Coastal Vulnerability
     * Improved handling of invalid AOI geometries to avoid crashing and instead
       fix the geometry when possible and skip it otherwise.
+    * Added validation check that shows a warning if the SLR vector is not 
+      a point or multipoint geometry.
 * Urban Cooling
     * Energy units are now (correctly) expressed in kWh.  They were previously
       (incorrectly) expressed in kW.
@@ -101,6 +107,9 @@ Unreleased Changes
 * Recreation
     * Validate values in the type column of predictor tables early in execution. Raise
       a ValueError if a type value isn't valid (leading/trailing whitespace is okay).
+* Validation
+    * Set a 5-second timeout on validation functions that access a file. This will raise 
+      a warning and prevent validation from slowing down the UI too much.
 
 3.8.7 (2020-07-17)
 ------------------